refactor(base): returning ResponseEntity<Mono<O>>
